move activity generations to stardust folder
