+++ packages/shared/src/lib/contexts/governance/actions/setVotingPower.ts
@@ -1,7 +1,7 @@
 import { PreparedTransaction } from '@iota/sdk/out/types'
 import { getSelectedAccount, updateSelectedAccount } from '@core/account/stores'
 import { handleError } from '@core/error/handlers'
-import { processAndAddToActivities } from '@core/activity/utils/processAndAddToActivities'
+import { processAndAddToActivities } from '@core/activity/actions/processAndAddToActivities'
 import { sendPreparedTransaction } from '@core/wallet'
 import { localize } from '@core/i18n'
 import { getActiveNetworkId } from '@core/network'
