+++ packages/shared/src/lib/core/activity/actions/index.ts
@@ -9,6 +9,7 @@ export * from './linkTransactionsWithClaimingTransactions'
 export * from './preprocessOutputsForAccount'
 export * from './preprocessTransactionsForAccount'
 export * from './loadAssetsForAllAccounts'
+export * from './processAndAddToActivities'
 export * from './setAsyncStatusOfAccountActivities'
 export * from './setOutgoingAsyncActivitiesToClaimed'
 export * from './updateClaimingTransactionInclusion'
