+++ packages/shared/src/lib/core/activity/actions/processAndAddToActivities.ts
@@ -2,8 +2,8 @@ import type { Transaction } from '@iota/sdk/out/types'
 
 import type { IAccountState } from '@core/account/interfaces'
 import { addAccountActivities } from '../stores'
-import { preprocessTransaction } from './outputs'
-import { generateActivities } from './generateActivities'
+import { preprocessTransaction } from '../utils/outputs'
+import { generateActivitiesFromStardustNetwork } from '../utils/stardust/generateActivitiesFromStardustNetwork'
 import { NetworkId } from '@core/network/types'
 
 // We pass the account as a parameter,
@@ -15,6 +15,6 @@ export async function processAndAddToActivities(
     networkId: NetworkId
 ): Promise<void> {
     const preprocessedTransaction = await preprocessTransaction(transaction, account)
-    const activities = await generateActivities(preprocessedTransaction, account, networkId)
+    const activities = await generateActivitiesFromStardustNetwork(preprocessedTransaction, account, networkId)
     addAccountActivities(account.index, activities)
 }
