+++ packages/shared/src/lib/core/activity/utils/helper/getSendingInformation.ts
@@ -2,8 +2,8 @@ import { IAccountState } from '@core/account'
 import { Output } from '@core/wallet/types'
 import { getRecipientFromOutput } from '../outputs'
 import { getSubjectFromAddress, isSubjectInternal } from '@core/wallet/utils'
-import { getSenderAddressFromInputs } from '../getSenderAddressFromInputs'
-import { getSenderFromTransaction } from '../getSenderFromTransaction'
+import { getSenderAddressFromInputs } from '../stardust/getSenderAddressFromInputs'
+import { getSenderFromTransaction } from '../stardust/getSenderFromTransaction'
 import { ActivityDirection } from '@core/activity/enums'
 import { IProcessedTransaction, SenderInfo } from '@core/activity/types'
 import { NetworkId } from '@core/network'
