+++ packages/shared/src/lib/core/activity/utils/index.ts
@@ -1,32 +1,13 @@
-export * from './activityOutputContainsValue'
-export * from './generateActivities'
-export * from './generateActivitiesFromAliasOutputs'
 export * from './generateActivitiesFromBalanceChanges'
-export * from './generateActivitiesFromBasicOutputs'
-export * from './generateActivitiesFromChains'
-export * from './generateActivitiesFromFoundryOutputs'
-export * from './generateActivitiesFromNftOutputs'
-export * from './generateActivitiesFromProcessedTransactions'
-export * from './evm/generateNftBalanceChangeActivity'
-export * from './generateSingleBasicActivity'
-export * from './generateSingleConsolidationActivity'
-export * from './generateSingleFoundryActivity'
-export * from './generateSingleGovernanceActivity'
-export * from './generateSingleNftActivity'
 export * from './getActivityDetailsTitle'
 export * from './getActivityActionTextColor'
 export * from './getActivityActionPill'
 export * from './getActivityTileAction'
 export * from './getActivityTileAsset'
-export * from './getDirectionFromTransaction'
-export * from './getNonRemainderBasicOutputsFromTransaction'
-export * from './getSenderAddressFromInputs'
-export * from './getSenderFromTransaction'
 export * from './getTransactionAssets'
-export * from './helper'
-export * from './isOutputSubjectFromActiveAccount'
 export * from './isVisibleActivity'
-export * from './processAndAddToActivities'
 
 export * from './evm'
+export * from './helper'
 export * from './outputs'
+export * from './stardust'
