+++ packages/shared/src/lib/core/activity/utils/stardust/generateActivitiesFromBasicOutputs.ts
@@ -2,13 +2,13 @@ import { IAccountState } from '@core/account'
 import { updateAllAccountNftsForAccount, buildNftFromNftOutput } from '@core/nfts/actions'
 import { IWrappedOutput } from '@core/wallet'
 import { NftOutput, OutputType } from '@iota/sdk/out/types'
-import { ActivityAction, ActivityDirection } from '../enums'
-import { StardustActivity, IProcessedTransaction } from '../types'
+import { ActivityAction, ActivityDirection } from '../../enums'
+import { StardustActivity, IProcessedTransaction } from '../../types'
 import { generateSingleBasicActivity } from './generateSingleBasicActivity'
 import { generateSingleConsolidationActivity } from './generateSingleConsolidationActivity'
 import { generateSingleNftActivity } from './generateSingleNftActivity'
 import { getNonRemainderBasicOutputsFromTransaction } from './getNonRemainderBasicOutputsFromTransaction'
-import { getNftId } from './outputs'
+import { getNftId } from '../outputs'
 import { NetworkId } from '@core/network/types'
 
 export async function generateActivitiesFromBasicOutputs(
