+++ packages/shared/src/lib/core/activity/utils/stardust/generateActivitiesFromFoundryOutputs.ts
@@ -1,8 +1,8 @@
 import { OutputType } from '@iota/sdk/out/types'
 import { IAccountState } from '@core/account'
-import { StardustActivity, IProcessedTransaction } from '../types'
+import { StardustActivity, IProcessedTransaction } from '../../types'
 import { generateSingleFoundryActivity } from './generateSingleFoundryActivity'
-import { ActivityAction } from '../enums'
+import { ActivityAction } from '../../enums'
 import { NetworkId } from '@core/network/types'
 
 export async function generateActivitiesFromFoundryOutputs(
