+++ packages/shared/src/lib/core/activity/utils/stardust/generateActivitiesFromNftOutputs.ts
@@ -1,9 +1,9 @@
 import { IAccountState } from '@core/account'
 import { EMPTY_HEX_ID } from '@core/wallet'
-import { StardustActivity, IProcessedTransaction } from '../types'
+import { StardustActivity, IProcessedTransaction } from '../../types'
 import { OutputType, NftOutput } from '@iota/sdk/out/types'
 import { generateSingleNftActivity } from './generateSingleNftActivity'
-import { ActivityAction } from '../enums'
+import { ActivityAction } from '../../enums'
 import { NetworkId } from '@core/network/types'
 
 export async function generateActivitiesFromNftOutputs(
