+++ packages/shared/src/lib/core/activity/utils/stardust/generateSingleBasicActivity.ts
@@ -2,8 +2,8 @@ import { IAccountState } from '@core/account'
 import { IActivityGenerationParameters } from '@core/activity/types'
 import { NetworkId } from '@core/network/types'
 import { BASE_TOKEN_ID } from '@core/token'
-import { StardustActivityType } from '../enums'
-import { StardustTransactionActivity } from '../types'
+import { StardustActivityType } from '../../enums'
+import { StardustTransactionActivity } from '../../types'
 import { generateBaseActivity } from './generateBaseActivity'
 import { getOrRequestTokenFromPersistedTokens } from '@core/token/actions'
