+++ packages/shared/src/lib/core/activity/utils/stardust/generateSingleGovernanceActivity.ts
@@ -1,11 +1,11 @@
 import { IAccountState } from '@core/account'
 import type { BasicOutput } from '@iota/sdk/out/types'
-import {} from '../types'
-import { StardustGovernanceActivity, IActivityGenerationParameters } from '../types'
-import { StardustActivityType } from '../enums'
+import {} from '../../types'
+import { StardustGovernanceActivity, IActivityGenerationParameters } from '../../types'
+import { StardustActivityType } from '../../enums'
 import { NetworkId } from '@core/network/types'
 import { generateBaseActivity } from './generateBaseActivity'
-import { getGovernanceInfo } from './helper'
+import { getGovernanceInfo } from '../helper'
 
 export async function generateSingleGovernanceActivity(
     account: IAccountState,
