+++ packages/shared/src/lib/core/activity/utils/stardust/generateSingleNftActivity.ts
@@ -1,11 +1,11 @@
 import { IAccountState } from '@core/account'
 import type { NftOutput } from '@iota/sdk/out/types'
-import { ActivityAction, StardustActivityType } from '../enums'
-import { StardustNftActivity } from '../types'
+import { ActivityAction, StardustActivityType } from '../../enums'
+import { StardustNftActivity } from '../../types'
 import { NetworkId } from '@core/network/types'
-import { IActivityGenerationParameters } from '../types/activity-generation-parameters.interface'
+import { IActivityGenerationParameters } from '../../types/activity-generation-parameters.interface'
 import { generateSingleBasicActivity } from './generateSingleBasicActivity'
-import { getNftId } from './outputs'
+import { getNftId } from '../outputs'
 import { BASE_TOKEN_ID } from '@core/token/constants'
 
 export async function generateSingleNftActivity(
