+++ packages/shared/src/lib/core/activity/utils/stardust/getSenderAddressFromInputs.ts
