+++ packages/shared/src/lib/core/activity/utils/stardust/isOutputSubjectFromActiveAccount.ts
@@ -3,7 +3,7 @@ import { NetworkId } from '@core/network'
 import { SubjectType } from '@core/wallet/enums'
 import { getSubjectFromAddress } from '@core/wallet/utils'
 import { CommonOutput } from '@iota/sdk/out/types'
-import { getRecipientAddressFromOutput } from './outputs'
+import { getRecipientAddressFromOutput } from '../outputs'
 
 export function isOutputSubjectFromActiveAccount(
     output: CommonOutput | undefined,
