+++ packages/shared/src/lib/core/profile-manager/actions/events-handlers/handleNewOutputEvent.ts
@@ -5,7 +5,7 @@ import { checkAndRemoveProfilePicture } from '@core/profile/actions'
 import { activeAccounts } from '@core/profile/stores'
 import { IWrappedOutput } from '@core/wallet/interfaces'
 import { addAccountActivities, allAccountActivities } from '@core/activity/stores/all-account-activities.store'
-import { generateActivities } from '@core/activity/utils'
+import { generateActivitiesFromStardustNetwork } from '@core/activity/utils'
 import { preprocessGroupedOutputs } from '@core/activity/utils/outputs'
 import { getActiveNetworkId } from '@core/network'
 import { addNftsToDownloadQueue, updateAllAccountNftsForAccount, buildNftFromNftOutput } from '@core/nfts/actions'
@@ -45,7 +45,7 @@ export async function handleNewOutputEventInternal(
 
         const processedOutput = preprocessGroupedOutputs([output], walletEvent?.transactionInputs ?? [], account)
 
-        const activities = await generateActivities(processedOutput, account, networkId)
+        const activities = await generateActivitiesFromStardustNetwork(processedOutput, account, networkId)
         for (const activity of activities) {
             if (activity.type === StardustActivityType.Basic || activity.type === StardustActivityType.Foundry) {
                 const tokenId = activity.tokenTransfer?.tokenId ?? activity.baseTokenTransfer?.tokenId
