+++ packages/shared/src/lib/core/wallet/actions/mintNft.ts
@@ -4,7 +4,7 @@ import { ActivityAction } from '@core/activity/enums'
 import { addAccountActivity } from '@core/activity/stores'
 import { sendPreparedTransaction } from '@core/wallet/utils'
 import { StardustNftActivity } from '@core/activity/types'
-import { generateSingleNftActivity } from '@core/activity/utils/generateSingleNftActivity'
+import { generateSingleNftActivity } from '@core/activity/utils/stardust/generateSingleNftActivity'
 import { preprocessTransaction } from '@core/activity/utils/outputs'
 import { localize } from '@core/i18n'
 import { IIrc27Metadata } from '@core/nfts'
